Fix rate limit of 600 requests per minutes on Gitlab.com (#106)
This is a simple fix, maximum of 10 requests per seconds = maximum of 600 requests per minutes. It does slow down a bit the report but it prevent a fatal error. It's possible to design a more sophisticated fix but it will be more complex with more states to manage. It would be more prone to bugs. I favor simplicity over maximum speed.

Co-authored-by: Jean-Denis Caron <jd.caron@xyz.zone>
Co-authored-by: Ciprian Dobre-Trifan <cgdobre@gmail.com>